feat: improved ui for launch screen, added react navigation for navigating between screens
