+++ GameScreen.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import {
+    Text,
+  View,
+} from 'react-native';
+
+
+function GameScreen() {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Text>Home Screen</Text>
+      </View>
+    );
+  }
+
+export default GameScreen;
