+++ package.json
@@ -10,9 +10,17 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.8",
+    "@react-navigation/stack": "^6.3.18",
     "react": "18.2.0",
     "react-native": "0.72.4",
-    "react-native-linear-gradient": "^2.8.3"
+    "react-native-gesture-handler": "^2.13.3",
+    "react-native-linear-gradient": "^2.8.3",
+    "react-native-reanimated": "^3.5.4",
+    "react-native-safe-area-context": "^4.7.3",
+    "react-native-screens": "^3.26.0",
+    "react-navigation": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
